Refactor tab drop delegate
